6-size.c correction

+++ 0x00-hello_world/6-size.c
@@ -7,13 +7,13 @@
  */
 int main(void)
 {
-printf("Size of char: %u byte(s)\n", (unsigned int)sizeof(char));
-printf("Size of int: %u byte(s)\n", (unsigned int)sizeof(int));
-printf("Size of long: %u byte(s)\n", (unsigned int)sizeof(long));
-printf("Size of long long: %u byte(s)\n", (unsigned int)sizeof(long long));
-printf("Size of float: %u byte(s)\n", (unsigned int)sizeof(float));
-printf("Size of double: %u byte(s)\n", (unsigned int)sizeof(double));
-printf("Size of long double: %u byte(s)\n", (unsigned int)sizeof(long double));
+printf("Size of char: %d byte(s)\n", sizeof(char));
+printf("Size of int: %d byte(s)\n", sizeof(int));
+printf("Size of long: %d byte(s)\n", sizeof(long));
+printf("Size of long long: %d byte(s)\n", sizeof(long long));
+printf("Size of float: %d byte(s)\n", sizeof(float));
+printf("Size of double: %d byte(s)\n", sizeof(double));
+printf("Size of long double: %d byte(s)\n", sizeof(long double));
 
 return (0);
 }
